ci: fix paths in parallel execution tests

+++ .github/compare_parallel_results.sh
@@ -1,24 +1,36 @@
 #! /bin/bash
 
+zarr1=bin_1/fromfile_sol.zarr
+if [ ! -d $zarr1 ]; then
+  echo "$zarr1 does not exist."
+  exit 1
+fi
+
 for processes in 2 4 8; do
     if [ -s "diffs" ]; then
         rm diffs
     fi
 
-    for file in $(ls bin_1/fromfile_sol.zarr/); do
-        filename=$(ls bin_1/fromfile_sol.zarr/$file)
-        hexdump bin_1/fromfile_sol.zarr/$file/$filename > original
-        hexdump bin_${processes}/fromfile_sol.zarr/$file/$filename > new
+    zarrX=bin_${processes}/fromfile_sol.zarr
+    if [ ! -d $zarrX ]; then
+        echo "$zarrX does not exist."
+        exit 1
+    fi
+
+    for file in $(ls $zarr1/); do
+        filename=$(ls $zarr1/$file)
+        hexdump $zarr1/$file/$filename > original
+        hexdump $zarrX/$file/$filename > new
         diff original new >> diffs
     done;
     rm new original
 
-    diff bin_1/fromfile_sol.zarr/.zattrs bin_${processes}/fromfile_sol.zarr/.zattrs >> diffs
-    diff bin_1/fromfile_sol.zarr/.zgroup bin_${processes}/fromfile_sol.zarr/.zgroup >> diffs
-    for file in $(ls bin_1/fromfile_sol.zarr/); do
-        filename=$(ls -la bin_1/fromfile_sol.zarr/$file)
-        diff bin_1/fromfile_sol.zarr/$file/.zattrs bin_${processes}/fromfile_sol.zarr/$file/.zattrs >> diffs
-        diff bin_1/fromfile_sol.zarr/$file/.zarray bin_${processes}/fromfile_sol.zarr/$file/.zarray >> diffs
+    diff $zarr1/.zattrs $zarrX/.zattrs >> diffs
+    diff $zarr1/.zgroup $zarrX/.zgroup >> diffs
+    for file in $(ls $zarr1/); do
+        filename=$(ls -la $zarr1/$file)
+        diff $zarr1/$file/.zattrs $zarrX/$file/.zattrs >> diffs
+        diff $zarr1/$file/.zarray $zarrX/$file/.zarray >> diffs
     done;
 
     if [ -s "diffs" ]; then
